requirements_sd_app.txt
flask
python-docx
ollama
werkzeug
openpyxl
